backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pyjwt
python-multipart
httpx
pydantic-settings
# Supabase sub-packages (installed individually to avoid storage3/pyiceberg C++ build issues)
postgrest
supabase-auth
supabase-functions
realtime
